refactor(tag): move some methods in the useAddTag file to the new file for unified management
