+++ src/composables/config/useAddTag.ts
@@ -1,170 +1,17 @@
 import { addTag, queryPluginConfigInfo } from '@/api/config'
 import type { PluginInfo, TagFormItem, AddTagListForm, TagForm } from '@/types/config'
-import { DriverDirection, TagAttributeType, TagType } from '@/types/enums'
+import { DriverDirection } from '@/types/enums'
 import type { Ref } from 'vue'
 import { ref, computed } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { EmqxMessage } from '@emqx/emqx-ui'
 import { useI18n } from 'vue-i18n'
 import type TagFormCom from '@/views/config/southDriver/components/TagForm.vue'
-import { createRandomString, dataType, getErrorMsg, popUpErrorMessage } from '@/utils/utils'
+import { dataType, getErrorMsg, popUpErrorMessage } from '@/utils/utils'
 import { useNodeMsgMap } from './useNodeList'
 import { useGetPluginMsgIdMap } from './usePlugin'
 import useWriteDataCheckNParse from '@/composables/data/useWriteDataCheckNParse'
-
-export const useTagTypeSelect = () => {
-  const tagTypeOptList = Object.keys(TagType)
-    .filter((key) => typeof TagType[key as keyof typeof TagType] === 'string')
-    .map((key) => ({
-      value: Number(key),
-      label: TagType[key as keyof typeof TagType],
-    }))
-
-  const findLabelByValue = (val: number) => tagTypeOptList.find(({ value }) => val === value)?.label || ''
-
-  const findValueByLabel = (lab: TagType) => tagTypeOptList.find(({ label }) => label === lab)?.value || undefined
-
-  return {
-    tagTypeOptList,
-    findLabelByValue,
-    findValueByLabel,
-  }
-}
-
-export const useTagAttributeTypeSelect = () => {
-  const tagAttributeTypeOptList = Object.keys(TagAttributeType)
-    .filter((key) => typeof TagAttributeType[key as keyof typeof TagAttributeType] === 'string')
-    .map((key) => ({
-      value: Number(key),
-      label: TagAttributeType[key as keyof typeof TagAttributeType],
-    }))
-
-  /** Tag `attribute` possible value
-   *  Options: 1, 2, 4, 8
-   *  key-value: 'value' is selected values; 'key' is the cumulative of values;
-   *  For example, 6: [2, 4], means that the the currently selected values are 2, 4, then the `key` is the sum of 2 and 4, i.e. 6.
-   */
-  const tagAttrValueMap = {
-    1: [1],
-    2: [2],
-    3: [1, 2],
-    4: [4],
-    5: [1, 4],
-    6: [2, 4],
-    7: [1, 2, 4],
-    8: [8],
-    9: [1, 8],
-    10: [2, 8],
-    11: [1, 2, 8],
-    12: [4, 8],
-    13: [1, 4, 8],
-    14: [2, 4, 8],
-    15: [1, 2, 4, 8],
-  }
-
-  const findLabelByValue = (val: number) => tagAttributeTypeOptList.find(({ value }) => val === value)?.label || ''
-  const findValueByLabel = (lab: any) => tagAttributeTypeOptList.find(({ label }) => label === lab)?.value || undefined
-
-  const getAttrStrByValue = (val: number, filler = ', ') => {
-    const key = Number(Object.keys(tagAttrValueMap).find((total) => Number(total) === val))
-    const valueArr = tagAttrValueMap[key as keyof typeof tagAttrValueMap]
-    return valueArr.map((value) => findLabelByValue(value)).join(filler)
-  }
-
-  // "Write Subscribe" to 5
-  const getTotalValueByStr = (str: string, separator: string): number | undefined => {
-    try {
-      if (!str) {
-        return undefined
-      }
-      const labelArr = str.split(separator)
-      let ret = 0
-      let currentValue: number | undefined = 0
-      for (const label of labelArr) {
-        currentValue = findValueByLabel(label)
-        if (!currentValue) {
-          return undefined
-        }
-        ret += currentValue
-      }
-      return ret
-    } catch (error) {
-      return undefined
-    }
-  }
-
-  // attr values is include the attr value
-  const isAttrsIncludeTheValue = (value: number | undefined, attr: number) => {
-    if (!value || !attr) return false
-
-    const key = Object.keys(tagAttrValueMap).find((total) => {
-      return Number(total) === value
-    })
-    const attrValues = tagAttrValueMap[Number(key) as keyof typeof tagAttrValueMap]
-    return attrValues.includes(attr)
-  }
-
-  return {
-    tagAttributeTypeOptList,
-    tagAttrValueMap,
-    getAttrStrByValue,
-    findLabelByValue,
-    getTotalValueByStr,
-    isAttrsIncludeTheValue,
-  }
-}
-
-export const useTagPrecision = () => {
-  const isShowPrecisionField = computed(() => (type: number | null) => {
-    if (type === null || type === undefined) return false
-
-    const whiteList = [9, 10] // FLOAT | DOUBLE
-    const res: boolean = whiteList.includes(type)
-    return res
-  })
-
-  const tagPrecisionValue = computed(() => (type: number, value: number) => {
-    if (!isShowPrecisionField.value(type)) {
-      return '-'
-    }
-    return !value ? '-' : value
-  })
-
-  return {
-    isShowPrecisionField,
-    tagPrecisionValue,
-  }
-}
-export const useTagDecimal = () => {
-  const tagDecimalValue = computed(() => (value: number) => {
-    if (!value) {
-      return '-'
-    }
-    return value
-  })
-
-  return {
-    tagDecimalValue,
-  }
-}
-export const createTagForm = () => {
-  const createRawTagForm = (): TagFormItem => ({
-    name: '',
-    address: '',
-    attribute: undefined,
-    type: null,
-    // for the key when use v-for
-    id: createRandomString(6),
-    decimal: null,
-    description: '',
-    precision: undefined,
-    value: undefined,
-  })
-
-  return {
-    createRawTagForm,
-  }
-}
+import { createTagForm } from '@/composables/config/useAddTagCommon'
 
 export default () => {
   const route = useRoute()
