+++ src/views/config/southDriver/components/TagListForm.vue
@@ -143,7 +143,7 @@
 import type { PropType, WritableComputedRef } from 'vue'
 import { defineExpose, computed, defineProps, defineEmits, nextTick } from 'vue'
 import { ElTable, ElForm, ElFormItem, ElTableColumn } from 'element-plus'
-import { useTagPrecision } from '@/composables/config/useAddTag'
+import { useTagPrecision } from '@/composables/config/useAddTagCommon'
 import TagAttributeSelect from './TagAttributeSelect.vue'
 import type { PluginInfo, AddTagListForm, TagFormItem } from '@/types/config'
 import useTagForm from '@/composables/config/useTagForm'
