+++ src/views/monitoring/DataMonitoring.vue
@@ -124,7 +124,7 @@ import { ref } from 'vue'
 import { ElPopover } from 'element-plus'
 import type { TagDataInTable } from '@/composables/data/useDataMonitoring'
 import useDataMonitoring from '@/composables/data/useDataMonitoring'
-import { useTagDecimal } from '@/composables/config/useAddTag'
+import { useTagDecimal } from '@/composables/config/useAddTagCommon'
 import dateformat from 'dateformat'
 import { getErrorMsg } from '@/utils/utils'
 import WriteDialog from './components/WriteDialog.vue'
